% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots FARS accident locations on the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer that represents unique code}

\item{year}{numeric value specifying year}
}
\value{
Returns \code{invisible(NULL)} and gives a geographical plot
}
\description{
Function will filter for specific state and year
}
\examples{
\dontrun{
fars_map_state(1, 2013)
fars_map_state(36. 2015)
}

}
