% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Week 2 Functions Explanation}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string providing path and name of CSV file}
}
\value{
Decription of output
Read FARS CSV file

Function tries to reach a CSV file about National Highway Safety
Administration data and loads it into R session as \code{tibble} or data
frame

a data frame referring to \code{tibble} that contains FARS data
Give an error message "file filename does not exist"
}
\description{
Week 2 Functions Explanation
}
\examples{
\dontrun{
fars_read("accident_2014.csv.bz2")
}
#' @export
}
