% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS data for multiple years at a time}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{numeric vector, list, or array of years}
}
\value{
A list of data frames containing the MONTH and year
columns \code{NULL} for list item
gives message saying "invalid year:year"
}
\description{
Function takes a vector, creates the corresponding FARS filenames, and uploads
data for the years.
}
\examples{
\dontrun{
fars_read_years(c(2014, 2015))
}

}
