% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes FARS data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{is a numeric vector of years to be summarized}
}
\value{
A wide data frame (\code{tibble}) that summarizes number of accidents
for that year (columns)
}
\description{
Summarizes FARS data by year and month
}
\examples{
\dontrun{
fars_summarize_years(c(2014, 2015))}

}
