% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a standardized FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{provides a numeric value that represents the target year. It can
convert other formats into an integer.}
}
\value{
Returns a character string formatted as
\code{"accident_[year].csv.bz2"}
}
\description{
Function creates a character string that shows the filename for a FARS Data
frame file giving the year. Example= \code{"accident_2015.csv.bz2"}.
}
\examples{
make_filename(2015)
make_filename(2014)

}
