% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge.R
\name{opt_ridge}
\alias{opt_ridge}
\title{Find the optimized value of parameter lambda in ridge regression}
\usage{
opt_ridge(X, y, lambda_vals)
}
\arguments{
\item{X}{A data frame containing the variables in the model. The levels of categorical variables should be
represented as numeric.}

\item{y}{A vector of the quantitative response variable.}

\item{lambda_vals}{A vector of possible values of the tuning paramenter \eqn{\lambda} in ridge regression.}
}
\value{
\code{opt_ridge} returns the optimized value of parameter \eqn{\lambda}.
  And a plot of mean squared errors versus \eqn{\lambda} is produced.
}
\description{
\code{opt_ridge} returns the optimized ridge parameter \eqn{\lambda} within a given set of possible values of \eqn{\lambda}.
  Cross validation is used and optimization is achieved by the least value of mean squared error.
}
\examples{
data(iris)
X <- iris[,2:5]
X$Species <- as.numeric(X$Species)
y <- iris[,1]
lambda <- opt_ridge(X, y, 1:500)
}
\keyword{regression}
\keyword{ridge}
